Adding 4 methods
Deduce[Some]MapFromMatrixAnd(Source/Range)ForGraded(Rows/Cols)

In the current state the map is returned only if it is unique determined.

The same code with small edits can be used to return some map whenever
the map is not uniquely determined.

Choosing the map happens like this:
Whenever a degree is not uniquely determined, choose Degree( One( S ) ).