Handle fallbacks for IsEqualForObjects and IsEqualForMorphisms in a better way
Currently, IsEqualForObjects and IsEqualForMorphisms fall back to
ReturnFail via derivations, with the redirect functions checking
IsIdenticalObj.
This has the following advantages:
1. IsEqualForObjects/Morphisms are always available.
2. As long as things are identical, everything "just works".
3. If things are not identical, one gets a more or less clear error message
   in pre-functions.

It has the following disadvantages:
1. Returning `fail` is not covered by the specification.
2. Every pre-function has to explicitly take `fail` into account.
3. Derivations like IsEndomorphism get triggered although they are not
   computable.
4. The fallback is not compatible with the specification "equal input gives
   equal output" if strict caching is not enabled manually.

This commit implements a better solution:
Instead of relying on fallback derivations for IsEqualForObjects and
IsEqualForMorphisms, we rely on the existing global plausibility checks
for IsEqualForObjects and IsEqualForMorphisms. Hence, no other derivations
are triggered and the specification "equal input gives equal output" simply
has no meaning. Instead of returning `fail` an error message is displayed.
Thus, pre-functions do not have to take `fail` into account. Also, the
error messages are more helpful then the ones in the pre-functions were.